fix: clear quota fallback CI gates
