+++ open-sse/services/combo.ts
@@ -2350,19 +2350,12 @@ export async function handleComboChat({
             !isTokenLimitBreach &&
             [408, 429, 500, 502, 503, 504].includes(result.status);
           if (retry < maxRetries && isTransient && !providerExhausted) {
-            // The auth layer (handleSingleModelChat) may have already recorded a
-            // model lockout for this provider:model (e.g. on 500 server_error).
-            // Check before retrying — retrying a locked model wastes an upstream
-            // call and extends the cooldown. Let the combo try the next target.
             if (
               provider &&
               rawModel &&
               isModelLocked(provider, targetWithConnection.connectionId || "", rawModel)
             ) {
-              log.info(
-                "COMBO",
-                `Skipping retry for ${modelStr} — model locked by auth layer (cooldown active)`
-              );
+              log.info("COMBO", `Skipping retry for ${modelStr} — model lockout active`);
               if (i > 0) fallbackCount++;
               return null;
             }
